moved get user model function to Utils.php
